+++ host/fcapz/__init__.py
@@ -10,6 +10,7 @@ from .analyzer import (
     SequencerStage,
     TriggerConfig,
 )
+from .axi_decode import AxiTransaction, decode_axi, looks_like_axi
 from .axi_monitor import AxiGeometry, AxiMonitor, AxiMonitorError
 from .eio import EioController
 from .ejtagaxi import AXIError, EjtagAxiController
@@ -59,6 +60,9 @@ __all__ = [
     "find_bursts",
     "frequency_estimate",
     "summarize",
+    "AxiTransaction",
+    "decode_axi",
+    "looks_like_axi",
     "ProbeDefinition",
     "AxiMonitor",
     "AxiGeometry",